requirements.txt
# Nutrition App Requirements
# Python 3.11.11
# Generated from working environment

# Core Flask framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21
alembic==1.16.1

# Forms and validation
WTForms==3.1.1
email-validator==2.2.0

# Security
itsdangerous==2.1.2

# Template engine
Jinja2==3.1.2
MarkupSafe==2.1.3

# HTTP requests
requests==2.31.0

# Image processing
Pillow==11.2.1

# Environment variables
python-dotenv==1.0.1

# Date/time utilities
python-dateutil==2.9.0.post0

# CLI utilities
click==8.1.7

# Additional utilities
colorама==0.4.6
blinker==1.9.0
